Rev Deps, Upgrade to Fiber, and Replace JSON Inspector (#279)
This commit updates the project's dependencies to latest, including updating React and React-DOM to 16.0.0-beta-5.

This is a significant update because React 16.x is the new "React Fiber" release with a compatibility shim that deprecated several APIs. The only component that was incompatible with React 16.x was JSON-Inspector, which we already wanted to replace due to sluggish performance.

This PR creates a replacement Inspector component that maintains enough of the CSS class name structure to allow @shanberg's styles in jsonInspector.scss to keep working. The search bar now filters our rows instead of highlighting matching terms, and the search and selection handlers are now router-driven and tied into browser history.

Resolves #276
Resolves #277
Resolves #278