Merge Fabric Wide Dash (#398)
This commit brings work from the `fabric-wide-dash` into `development`. Because of some shifting priorities, this work was put on the back burner, leading to a hairy rebase effort. As of this commit, the dashboard support a mock of the interim Discovery Service API while in "Fabric Server" mode, but it retains the ability to deploy as before when a "Fabric Server" endpoint is not provided.

This will provide the basis of further work implementing against the actual Discovery Service API in coming sprints.

* feat: Mock the API of the interim discovery service
* chore: containerize setup for dashboard and service through proxy
* feat: Implements initial Fabric-wide navigation
* feat: Basic polling of metrics restored.
* chore: Refactor JumpState Hooks into Effects to improve grok-ability
* fix: Update unit tests with new API.
* refactor: Organize flat src/components files hierarchically according to runtime and responsibility